Docs: move more constants to Utilities
